remove usage of echo -e from Debian init script
The debian default "/bin/sh" is dash. The build in echo
implementation in dash does not support "-e", so it echos
a verbatim "-e" but the \t are none the less parsed and
represented as a tabulator.
In case someone changed "/bin/sh" to a none default shell,
it might result in varbatim "\t" in the output.

Alternative would be to call "/bin/echo -e" explicitly.